+++ CHANGELOG.md
@@ -7,13 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.3.0] - 2020-05-26
+### Removed
+- `session` in client create API and Listener
+
 ## [1.2.0] - 2020-05-22
 ### Added
 - `databaseSettings` parameter for `postSaveHook()` method
 - `newClientsDatabaseKey` setting for new clients database config
-
-### Removed
-- `session` in client create API and Listener
 
 ### Changed
 - `Client` Model now set the new clients databaseKey from config